add score/default view, change fields to Decimal,  reform css, improve existing table sorting

+++ pontoon/insights/models.py
@@ -88,21 +88,31 @@ class ProjectLocaleInsightsSnapshot(InsightsSnapshot):
 class LocaleHealthSnapshot(models.Model):
     locale = models.ForeignKey("base.Locale", on_delete=models.CASCADE)
     created_at = models.DateField()
-    completion = models.FloatField(default=0.0)
+    completion = models.DecimalField(max_digits=5, decimal_places=2, default=0)
     key_projects_enabled = models.PositiveIntegerField(default=0)
     active_managers = models.PositiveIntegerField(default=0)
     active_translators = models.PositiveIntegerField(default=0)
     active_contributors = models.PositiveIntegerField(default=0)
     all_contributors = models.PositiveIntegerField(default=0)
     new_signups = models.PositiveIntegerField(default=0)
-    completion_score = models.FloatField(default=0.0)
-    key_projects_enabled_score = models.FloatField(default=0)
-    active_managers_score = models.FloatField(default=0)
-    active_translators_score = models.FloatField(default=0)
-    active_contributors_score = models.FloatField(default=0)
-    all_contributors_score = models.FloatField(default=0)
-    new_signups_score = models.FloatField(default=0)
-    chs = models.FloatField(default=0.0)
+    completion_score = models.DecimalField(max_digits=5, decimal_places=2, default=0)
+    key_projects_enabled_score = models.DecimalField(
+        max_digits=5, decimal_places=2, default=0
+    )
+    active_managers_score = models.DecimalField(
+        max_digits=5, decimal_places=2, default=0
+    )
+    active_translators_score = models.DecimalField(
+        max_digits=5, decimal_places=2, default=0
+    )
+    active_contributors_score = models.DecimalField(
+        max_digits=5, decimal_places=2, default=0
+    )
+    all_contributors_score = models.DecimalField(
+        max_digits=5, decimal_places=2, default=0
+    )
+    new_signups_score = models.DecimalField(max_digits=5, decimal_places=2, default=0)
+    chs = models.DecimalField(max_digits=5, decimal_places=2, default=0)
 
     class Meta:
         unique_together = [("locale", "created_at")]
